Cleaned up and refactored binary operator
