+++ src/expressions/operators/arithmetic/BinaryOperator.ts
@@ -1,29 +1,31 @@
-import { BinaryEvaluationFunction } from '../../../typeInference/binaryEvaluationFunction';
-import atomize from '../../dataTypes/atomize';
-import castToType from '../../dataTypes/castToType';
-import createAtomicValue from '../../dataTypes/createAtomicValue';
-import isSubtypeOf from '../../dataTypes/isSubtypeOf';
-import sequenceFactory from '../../dataTypes/sequenceFactory';
-import { SequenceType, ValueType } from '../../dataTypes/Value';
+import AtomicValue from '../../../expressions/dataTypes/AtomicValue';
+import castToType from '../../../expressions/dataTypes/castToType';
+import createAtomicValue from '../../../expressions/dataTypes/createAtomicValue';
+import isSubtypeOf from '../../../expressions/dataTypes/isSubtypeOf';
+import { ValueType } from '../../../expressions/dataTypes/Value';
 import {
 	addDuration as addDurationToDateTime,
 	subtract as dateTimeSubtract,
 	subtractDuration as subtractDurationFromDateTime,
-} from '../../dataTypes/valueTypes/DateTime';
+} from '../../../expressions/dataTypes/valueTypes/DateTime';
 import {
 	add as dayTimeDurationAdd,
 	divide as dayTimeDurationDivide,
 	divideByDayTimeDuration as dayTimeDurationDivideByDayTimeDuration,
 	multiply as dayTimeDurationMultiply,
 	subtract as dayTimeDurationSubtract,
-} from '../../dataTypes/valueTypes/DayTimeDuration';
+} from '../../../expressions/dataTypes/valueTypes/DayTimeDuration';
 import {
 	add as yearMonthDurationAdd,
 	divide as yearMonthDurationDivide,
 	divideByYearMonthDuration as yearMonthDurationDivideByYearMonthDuration,
 	multiply as yearMonthDurationMultiply,
 	subtract as yearMonthDurationSubtract,
-} from '../../dataTypes/valueTypes/YearMonthDuration';
+} from '../../../expressions/dataTypes/valueTypes/YearMonthDuration';
+import { BinaryEvaluationFunction } from '../../../typeInference/binaryEvaluationFunction';
+import atomize from '../../dataTypes/atomize';
+import sequenceFactory from '../../dataTypes/sequenceFactory';
+import { SequenceType } from '../../dataTypes/Value';
 import Expression from '../../Expression';
 
 function determineReturnType(typeA: ValueType, typeB: ValueType): ValueType {
@@ -39,12 +41,185 @@ function determineReturnType(typeA: ValueType, typeB: ValueType): ValueType {
 	return ValueType.XSDOUBLE;
 }
 
+const allTypes = [
+	ValueType.XSNUMERIC,
+	ValueType.XSYEARMONTHDURATION,
+	ValueType.XSDAYTIMEDURATION,
+	ValueType.XSDATETIME,
+	ValueType.XSDATE,
+	ValueType.XSTIME,
+];
+
+function hash(left: ValueType, right: ValueType, op: string): number {
+	return (
+		((left as number) << 20) +
+		((right as number) << 12) +
+		(op.charCodeAt(0) << 8) +
+		op.charCodeAt(1)
+	);
+}
+
+type EvalFuncTable = {
+	[key: number]: [(a: any, b: any) => any, ValueType];
+};
+
+const ruleMap: EvalFuncTable = {
+	[hash(ValueType.XSNUMERIC, ValueType.XSNUMERIC, 'addOp')]: [(a, b) => a + b, undefined],
+	[hash(ValueType.XSNUMERIC, ValueType.XSNUMERIC, 'subtractOp')]: [(a, b) => a - b, undefined],
+	[hash(ValueType.XSNUMERIC, ValueType.XSNUMERIC, 'multiplyOp')]: [(a, b) => a * b, undefined],
+	[hash(ValueType.XSNUMERIC, ValueType.XSNUMERIC, 'divOp')]: [(a, b) => a / b, undefined],
+	[hash(ValueType.XSNUMERIC, ValueType.XSNUMERIC, 'modOp')]: [(a, b) => a % b, undefined],
+	[hash(ValueType.XSNUMERIC, ValueType.XSNUMERIC, 'idivOp')]: [
+		(a, b) => Math.trunc(a / b),
+		ValueType.XSINTEGER,
+	],
+	[hash(ValueType.XSYEARMONTHDURATION, ValueType.XSYEARMONTHDURATION, 'addOp')]: [
+		yearMonthDurationAdd,
+		ValueType.XSYEARMONTHDURATION,
+	],
+	[hash(ValueType.XSYEARMONTHDURATION, ValueType.XSYEARMONTHDURATION, 'subtractOp')]: [
+		yearMonthDurationSubtract,
+		ValueType.XSYEARMONTHDURATION,
+	],
+	[hash(ValueType.XSYEARMONTHDURATION, ValueType.XSYEARMONTHDURATION, 'divOp')]: [
+		yearMonthDurationDivideByYearMonthDuration,
+		ValueType.XSDECIMAL,
+	],
+	[hash(ValueType.XSYEARMONTHDURATION, ValueType.XSNUMERIC, 'multiplyOp')]: [
+		yearMonthDurationMultiply,
+		ValueType.XSYEARMONTHDURATION,
+	],
+	[hash(ValueType.XSYEARMONTHDURATION, ValueType.XSNUMERIC, 'divOp')]: [
+		yearMonthDurationDivide,
+		ValueType.XSYEARMONTHDURATION,
+	],
+	[hash(ValueType.XSNUMERIC, ValueType.XSYEARMONTHDURATION, 'multiplyOp')]: [
+		(a, b) => yearMonthDurationMultiply(b, a),
+		ValueType.XSYEARMONTHDURATION,
+	],
+	[hash(ValueType.XSDAYTIMEDURATION, ValueType.XSDAYTIMEDURATION, 'addOp')]: [
+		dayTimeDurationAdd,
+		ValueType.XSDAYTIMEDURATION,
+	],
+	[hash(ValueType.XSDAYTIMEDURATION, ValueType.XSDAYTIMEDURATION, 'subtractOp')]: [
+		dayTimeDurationSubtract,
+		ValueType.XSDAYTIMEDURATION,
+	],
+	[hash(ValueType.XSDAYTIMEDURATION, ValueType.XSDAYTIMEDURATION, 'divOp')]: [
+		dayTimeDurationDivideByDayTimeDuration,
+		ValueType.XSDECIMAL,
+	],
+	[hash(ValueType.XSDAYTIMEDURATION, ValueType.XSNUMERIC, 'multiplyOp')]: [
+		dayTimeDurationMultiply,
+		ValueType.XSDAYTIMEDURATION,
+	],
+	[hash(ValueType.XSDAYTIMEDURATION, ValueType.XSNUMERIC, 'divOp')]: [
+		dayTimeDurationDivide,
+		ValueType.XSDAYTIMEDURATION,
+	],
+	[hash(ValueType.XSNUMERIC, ValueType.XSDAYTIMEDURATION, 'multiplyOp')]: [
+		(a, b) => dayTimeDurationMultiply(b, a),
+		ValueType.XSDAYTIMEDURATION,
+	],
+	[hash(ValueType.XSDATETIME, ValueType.XSDATETIME, 'subtractOp')]: [
+		dateTimeSubtract,
+		ValueType.XSDAYTIMEDURATION,
+	],
+	[hash(ValueType.XSDATE, ValueType.XSDATE, 'subtractOp')]: [
+		dateTimeSubtract,
+		ValueType.XSDAYTIMEDURATION,
+	],
+	[hash(ValueType.XSTIME, ValueType.XSTIME, 'subtractOp')]: [
+		dateTimeSubtract,
+		ValueType.XSDAYTIMEDURATION,
+	],
+	[hash(ValueType.XSDATETIME, ValueType.XSYEARMONTHDURATION, 'addOp')]: [
+		addDurationToDateTime,
+		ValueType.XSDATETIME,
+	],
+	[hash(ValueType.XSDATETIME, ValueType.XSYEARMONTHDURATION, 'subtractOp')]: [
+		subtractDurationFromDateTime,
+		ValueType.XSDATETIME,
+	],
+	[hash(ValueType.XSDATETIME, ValueType.XSDAYTIMEDURATION, 'addOp')]: [
+		addDurationToDateTime,
+		ValueType.XSDATETIME,
+	],
+	[hash(ValueType.XSDATETIME, ValueType.XSDAYTIMEDURATION, 'subtractOp')]: [
+		subtractDurationFromDateTime,
+		ValueType.XSDATETIME,
+	],
+	[hash(ValueType.XSDATE, ValueType.XSYEARMONTHDURATION, 'addOp')]: [
+		addDurationToDateTime,
+		ValueType.XSDATE,
+	],
+	[hash(ValueType.XSDATE, ValueType.XSYEARMONTHDURATION, 'subtractOp')]: [
+		subtractDurationFromDateTime,
+		ValueType.XSDATE,
+	],
+	[hash(ValueType.XSDATE, ValueType.XSDAYTIMEDURATION, 'addOp')]: [
+		addDurationToDateTime,
+		ValueType.XSDATE,
+	],
+	[hash(ValueType.XSDATE, ValueType.XSDAYTIMEDURATION, 'subtractOp')]: [
+		subtractDurationFromDateTime,
+		ValueType.XSDATE,
+	],
+	[hash(ValueType.XSTIME, ValueType.XSDAYTIMEDURATION, 'addOp')]: [
+		addDurationToDateTime,
+		ValueType.XSTIME,
+	],
+	[hash(ValueType.XSTIME, ValueType.XSDAYTIMEDURATION, 'subtractOp')]: [
+		subtractDurationFromDateTime,
+		ValueType.XSTIME,
+	],
+	[hash(ValueType.XSDATETIME, ValueType.XSYEARMONTHDURATION, 'addOp')]: [
+		addDurationToDateTime,
+		ValueType.XSDATETIME,
+	],
+	[hash(ValueType.XSDATETIME, ValueType.XSYEARMONTHDURATION, 'subtractOp')]: [
+		subtractDurationFromDateTime,
+		ValueType.XSDATETIME,
+	],
+	[hash(ValueType.XSDATETIME, ValueType.XSDAYTIMEDURATION, 'addOp')]: [
+		addDurationToDateTime,
+		ValueType.XSDATETIME,
+	],
+	[hash(ValueType.XSDATETIME, ValueType.XSDAYTIMEDURATION, 'subtractOp')]: [
+		subtractDurationFromDateTime,
+		ValueType.XSDATETIME,
+	],
+	[hash(ValueType.XSDATE, ValueType.XSDAYTIMEDURATION, 'addOp')]: [
+		addDurationToDateTime,
+		ValueType.XSDATE,
+	],
+	[hash(ValueType.XSDATE, ValueType.XSDAYTIMEDURATION, 'subtractOp')]: [
+		subtractDurationFromDateTime,
+		ValueType.XSDATE,
+	],
+	[hash(ValueType.XSDATE, ValueType.XSYEARMONTHDURATION, 'addOp')]: [
+		addDurationToDateTime,
+		ValueType.XSDATE,
+	],
+	[hash(ValueType.XSDATE, ValueType.XSYEARMONTHDURATION, 'subtractOp')]: [
+		subtractDurationFromDateTime,
+		ValueType.XSDATE,
+	],
+	[hash(ValueType.XSTIME, ValueType.XSDAYTIMEDURATION, 'addOp')]: [
+		addDurationToDateTime,
+		ValueType.XSTIME,
+	],
+	[hash(ValueType.XSTIME, ValueType.XSDAYTIMEDURATION, 'subtractOp')]: [
+		subtractDurationFromDateTime,
+		ValueType.XSTIME,
+	],
+};
+
 function generateBinaryOperatorFunction(
-	operator,
+	operator: string,
 	typeA: ValueType,
-	typeB: ValueType,
-	retType?: SequenceType
-) {
+	typeB: ValueType
+): [BinaryEvaluationFunction, ValueType] {
 	let castFunctionForValueA = null;
 	let castFunctionForValueB = null;
 
@@ -57,386 +232,104 @@ function generateBinaryOperatorFunction(
 		typeB = ValueType.XSDOUBLE;
 	}
 
-	function applyCastFunctions(valueA, valueB) {
+	const parentTypesOfA = allTypes.filter((e) => isSubtypeOf(typeA, e));
+	const parentTypesOfB = allTypes.filter((e) => isSubtypeOf(typeB, e));
+
+	function applyCastFunctions(valueA: AtomicValue, valueB: AtomicValue) {
 		return {
 			castA: castFunctionForValueA ? castFunctionForValueA(valueA) : valueA,
 			castB: castFunctionForValueB ? castFunctionForValueB(valueB) : valueB,
 		};
 	}
 
-	if (isSubtypeOf(typeA, ValueType.XSNUMERIC) && isSubtypeOf(typeB, ValueType.XSNUMERIC)) {
-		switch (operator) {
-			case 'addOp': {
-				const returnType = determineReturnType(typeA, typeB);
-				return (a, b) => {
-					const { castA, castB } = applyCastFunctions(a, b);
-
-					return createAtomicValue(castA.value + castB.value, returnType);
-				};
-			}
-			case 'subtractOp': {
-				const returnType = determineReturnType(typeA, typeB);
-				return (a, b) => {
-					const { castA, castB } = applyCastFunctions(a, b);
-					return createAtomicValue(castA.value - castB.value, returnType);
-				};
-			}
-			case 'multiplyOp': {
-				const returnType = determineReturnType(typeA, typeB);
-				return (a, b) => {
-					const { castA, castB } = applyCastFunctions(a, b);
-					return createAtomicValue(castA.value * castB.value, returnType);
-				};
-			}
-			case 'divOp': {
-				let returnType = determineReturnType(typeA, typeB);
-				if (returnType === ValueType.XSINTEGER) {
-					returnType = ValueType.XSDECIMAL;
-				}
-				return (a, b) => {
-					const { castA, castB } = applyCastFunctions(a, b);
-					return createAtomicValue(castA.value / castB.value, returnType);
-				};
-			}
-			case 'idivOp':
-				return (a, b) => {
-					const { castA, castB } = applyCastFunctions(a, b);
-					if (castB.value === 0) {
-						throw new Error('FOAR0001: Divisor of idiv operator cannot be (-)0');
-					}
-					if (
-						Number.isNaN(castA.value) ||
-						Number.isNaN(castB.value) ||
-						!Number.isFinite(castA.value)
-					) {
-						throw new Error(
-							'FOAR0002: One of the operands of idiv is NaN or the first operand is (-)INF'
-						);
-					}
-					if (Number.isFinite(castA.value) && !Number.isFinite(castB.value)) {
-						return createAtomicValue(0, ValueType.XSINTEGER);
-					}
-					return createAtomicValue(
-						Math.trunc(castA.value / castB.value),
-						ValueType.XSINTEGER
-					);
-				};
-			case 'modOp': {
-				const returnType = determineReturnType(typeA, typeB);
-
-				return (a, b) => {
-					const { castA, castB } = applyCastFunctions(a, b);
-					return createAtomicValue(castA.value % castB.value, returnType);
-				};
-			}
-		}
-	}
-
 	if (
-		isSubtypeOf(typeA, ValueType.XSYEARMONTHDURATION) &&
-		isSubtypeOf(typeB, ValueType.XSYEARMONTHDURATION)
+		parentTypesOfA.includes(ValueType.XSNUMERIC) &&
+		parentTypesOfB.includes(ValueType.XSNUMERIC)
 	) {
-		switch (operator) {
-			case 'addOp':
-				return (a, b) => {
-					const { castA, castB } = applyCastFunctions(a, b);
-					return createAtomicValue(
-						yearMonthDurationAdd(castA.value, castB.value),
-						ValueType.XSYEARMONTHDURATION
-					);
-				};
-			case 'subtractOp':
-				return (a, b) => {
-					const { castA, castB } = applyCastFunctions(a, b);
-					return createAtomicValue(
-						yearMonthDurationSubtract(castA.value, castB.value),
-						ValueType.XSYEARMONTHDURATION
-					);
-				};
-			case 'divOp':
-				return (a, b) => {
-					const { castA, castB } = applyCastFunctions(a, b);
-					return createAtomicValue(
-						yearMonthDurationDivideByYearMonthDuration(castA.value, castB.value),
-						ValueType.XSDECIMAL
-					);
-				};
-		}
-	}
-
-	if (
-		isSubtypeOf(typeA, ValueType.XSYEARMONTHDURATION) &&
-		isSubtypeOf(typeB, ValueType.XSNUMERIC)
-	) {
-		switch (operator) {
-			case 'multiplyOp':
-				return (a, b) => {
-					const { castA, castB } = applyCastFunctions(a, b);
-					return createAtomicValue(
-						yearMonthDurationMultiply(castA.value, castB.value),
-						ValueType.XSYEARMONTHDURATION
-					);
-				};
-			case 'divOp':
-				return (a, b) => {
-					const { castA, castB } = applyCastFunctions(a, b);
-					return createAtomicValue(
-						yearMonthDurationDivide(castA.value, castB.value),
-						ValueType.XSYEARMONTHDURATION
-					);
-				};
-		}
-	}
-
-	if (
-		isSubtypeOf(typeA, ValueType.XSNUMERIC) &&
-		isSubtypeOf(typeB, ValueType.XSYEARMONTHDURATION)
-	) {
-		if (operator === 'multiplyOp') {
-			return (a, b) => {
+		const result = ruleMap[hash(ValueType.XSNUMERIC, ValueType.XSNUMERIC, operator)];
+		const fun = result[0];
+		let retType = result[1];
+		if (!retType) retType = determineReturnType(typeA, typeB);
+		if (operator === 'divOp' && retType === ValueType.XSINTEGER) retType = ValueType.XSDECIMAL;
+		if (operator === 'idivOp') return idDivOpChecksFunction(applyCastFunctions, fun);
+		return [
+			(a, b) => {
 				const { castA, castB } = applyCastFunctions(a, b);
-				return createAtomicValue(
-					yearMonthDurationMultiply(castB.value, castA.value),
-					ValueType.XSYEARMONTHDURATION
-				);
-			};
-		}
+				return createAtomicValue(fun(castA.value, castB.value), retType);
+			},
+			retType,
+		];
 	}
-
-	if (
-		isSubtypeOf(typeA, ValueType.XSDAYTIMEDURATION) &&
-		isSubtypeOf(typeB, ValueType.XSDAYTIMEDURATION)
-	) {
-		switch (operator) {
-			case 'addOp':
-				return (a, b) => {
-					const { castA, castB } = applyCastFunctions(a, b);
-					return createAtomicValue(
-						dayTimeDurationAdd(castA.value, castB.value),
-						ValueType.XSDAYTIMEDURATION
-					);
-				};
-			case 'subtractOp':
-				return (a, b) => {
-					const { castA, castB } = applyCastFunctions(a, b);
-					return createAtomicValue(
-						dayTimeDurationSubtract(castA.value, castB.value),
-						ValueType.XSDAYTIMEDURATION
-					);
-				};
-			case 'divOp':
-				return (a, b) => {
-					const { castA, castB } = applyCastFunctions(a, b);
-					return createAtomicValue(
-						dayTimeDurationDivideByDayTimeDuration(castA.value, castB.value),
-						ValueType.XSDECIMAL
-					);
-				};
-		}
-	}
-	if (
-		isSubtypeOf(typeA, ValueType.XSDAYTIMEDURATION) &&
-		isSubtypeOf(typeB, ValueType.XSNUMERIC)
-	) {
-		switch (operator) {
-			case 'multiplyOp':
-				return (a, b) => {
-					const { castA, castB } = applyCastFunctions(a, b);
-					return createAtomicValue(
-						dayTimeDurationMultiply(castA.value, castB.value),
-						ValueType.XSDAYTIMEDURATION
-					);
-				};
-			case 'divOp':
-				return (a, b) => {
-					const { castA, castB } = applyCastFunctions(a, b);
-					return createAtomicValue(
-						dayTimeDurationDivide(castA.value, castB.value),
-						ValueType.XSDAYTIMEDURATION
-					);
-				};
-		}
-	}
-	if (
-		isSubtypeOf(typeA, ValueType.XSNUMERIC) &&
-		isSubtypeOf(typeB, ValueType.XSDAYTIMEDURATION)
-	) {
-		if (operator === 'multiplyOp') {
-			return (a, b) => {
-				const { castA, castB } = applyCastFunctions(a, b);
-				return createAtomicValue(
-					dayTimeDurationMultiply(castB.value, castA.value),
-					ValueType.XSDAYTIMEDURATION
-				);
-			};
-		}
-	}
-
-	if (
-		(isSubtypeOf(typeA, ValueType.XSDATETIME) && isSubtypeOf(typeB, ValueType.XSDATETIME)) ||
-		(isSubtypeOf(typeA, ValueType.XSDATE) && isSubtypeOf(typeB, ValueType.XSDATE)) ||
-		(isSubtypeOf(typeA, ValueType.XSTIME) && isSubtypeOf(typeB, ValueType.XSTIME))
-	) {
-		if (operator === 'subtractOp') {
-			return (a, b) => {
-				const { castA, castB } = applyCastFunctions(a, b);
-				return createAtomicValue(
-					dateTimeSubtract(castA.value, castB.value),
-					ValueType.XSDAYTIMEDURATION
-				);
-			};
-		}
-
+	if (parentTypesOfB.length === 0 || parentTypesOfA.length === 0)
 		throw new Error(`XPTY0004: ${operator} not available for types ${typeA} and ${typeB}`);
-	}
 
-	if (
-		(isSubtypeOf(typeA, ValueType.XSDATETIME) &&
-			isSubtypeOf(typeB, ValueType.XSYEARMONTHDURATION)) ||
-		(isSubtypeOf(typeA, ValueType.XSDATETIME) &&
-			isSubtypeOf(typeB, ValueType.XSDAYTIMEDURATION))
-	) {
-		if (operator === 'addOp') {
-			return (a, b) => {
-				const { castA, castB } = applyCastFunctions(a, b);
-				return createAtomicValue(
-					addDurationToDateTime(castA.value, castB.value),
-					ValueType.XSDATETIME
-				);
-			};
-		}
-		if (operator === 'subtractOp') {
-			return (a, b) => {
-				const { castA, castB } = applyCastFunctions(a, b);
-				return createAtomicValue(
-					subtractDurationFromDateTime(castA.value, castB.value),
-					ValueType.XSDATETIME
-				);
-			};
+	for (const typeOfA of parentTypesOfA) {
+		for (const typeOfB of parentTypesOfB) {
+			const result = ruleMap[hash(typeOfA, typeOfB, operator)];
+			if (result) {
+				return [
+					(a, b) => {
+						const { castA, castB } = applyCastFunctions(a, b);
+						return createAtomicValue(result[0](castA.value, castB.value), result[1]);
+					},
+					result[1],
+				];
+			}
 		}
 	}
-
-	if (
-		(isSubtypeOf(typeA, ValueType.XSDATE) &&
-			isSubtypeOf(typeB, ValueType.XSYEARMONTHDURATION)) ||
-		(isSubtypeOf(typeA, ValueType.XSDATE) && isSubtypeOf(typeB, ValueType.XSDAYTIMEDURATION))
-	) {
-		if (operator === 'addOp') {
-			return (a, b) => {
-				const { castA, castB } = applyCastFunctions(a, b);
-				return createAtomicValue(
-					addDurationToDateTime(castA.value, castB.value),
-					ValueType.XSDATE
-				);
-			};
-		}
-		if (operator === 'subtractOp') {
-			return (a, b) => {
-				const { castA, castB } = applyCastFunctions(a, b);
-				return createAtomicValue(
-					subtractDurationFromDateTime(castA.value, castB.value),
-					ValueType.XSDATE
-				);
-			};
-		}
-	}
-
-	if (isSubtypeOf(typeA, ValueType.XSTIME) && isSubtypeOf(typeB, ValueType.XSDAYTIMEDURATION)) {
-		if (operator === 'addOp') {
-			return (a, b) => {
-				const { castA, castB } = applyCastFunctions(a, b);
-				return createAtomicValue(
-					addDurationToDateTime(castA.value, castB.value),
-					ValueType.XSTIME
-				);
-			};
-		}
-		if (operator === 'subtractOp') {
-			return (a, b) => {
-				const { castA, castB } = applyCastFunctions(a, b);
-				return createAtomicValue(
-					subtractDurationFromDateTime(castA.value, castB.value),
-					ValueType.XSTIME
-				);
-			};
-		}
-	}
-
-	if (
-		(isSubtypeOf(typeB, ValueType.XSYEARMONTHDURATION) &&
-			isSubtypeOf(typeA, ValueType.XSDATETIME)) ||
-		(isSubtypeOf(typeB, ValueType.XSDAYTIMEDURATION) &&
-			isSubtypeOf(typeA, ValueType.XSDATETIME))
-	) {
-		if (operator === 'addOp') {
-			return (a, b) => {
-				const { castA, castB } = applyCastFunctions(a, b);
-				return createAtomicValue(
-					addDurationToDateTime(castB.value, castA.value),
-					ValueType.XSDATETIME
-				);
-			};
-		}
-		if (operator === 'subtractOp') {
-			return (a, b) => {
-				const { castA, castB } = applyCastFunctions(a, b);
-				return createAtomicValue(
-					subtractDurationFromDateTime(castB.value, castA.value),
-					ValueType.XSDATETIME
-				);
-			};
-		}
-	}
-
-	if (
-		(isSubtypeOf(typeB, ValueType.XSDAYTIMEDURATION) && isSubtypeOf(typeA, ValueType.XSDATE)) ||
-		(isSubtypeOf(typeB, ValueType.XSYEARMONTHDURATION) && isSubtypeOf(typeA, ValueType.XSDATE))
-	) {
-		if (operator === 'addOp') {
-			return (a, b) => {
-				const { castA, castB } = applyCastFunctions(a, b);
-				return createAtomicValue(
-					addDurationToDateTime(castB.value, castA.value),
-					ValueType.XSDATE
-				);
-			};
-		}
-		if (operator === 'subtractOp') {
-			return (a, b) => {
-				const { castA, castB } = applyCastFunctions(a, b);
-				return createAtomicValue(
-					subtractDurationFromDateTime(castB.value, castA.value),
-					ValueType.XSDATE
-				);
-			};
-		}
-	}
-
-	if (isSubtypeOf(typeB, ValueType.XSDAYTIMEDURATION) && isSubtypeOf(typeA, ValueType.XSTIME)) {
-		if (operator === 'addOp') {
-			return (a, b) => {
-				const { castA, castB } = applyCastFunctions(a, b);
-				return createAtomicValue(
-					addDurationToDateTime(castB.value, castA.value),
-					ValueType.XSTIME
-				);
-			};
-		}
-		if (operator === 'subtractOp') {
-			return (a, b) => {
-				const { castA, castB } = applyCastFunctions(a, b);
-				return createAtomicValue(
-					subtractDurationFromDateTime(castB.value, castA.value),
-					ValueType.XSTIME
-				);
-			};
-		}
-	}
-
 	throw new Error(`XPTY0004: ${operator} not available for types ${typeA} and ${typeB}`);
 }
 
-const operatorsByTypingKey = Object.create(null);
+function idDivOpChecksFunction(
+	applyCastFunctions: (a, b) => any,
+	fun: (a, b) => any
+): [(a: any, b: any) => AtomicValue, ValueType] {
+	return [
+		(a, b) => {
+			const { castA, castB } = applyCastFunctions(a, b);
+			if (castB.value === 0) {
+				throw new Error('FOAR0001: Divisor of idiv operator cannot be (-)0');
+			}
+			if (
+				Number.isNaN(castA.value) ||
+				Number.isNaN(castB.value) ||
+				!Number.isFinite(castA.value)
+			) {
+				throw new Error(
+					'FOAR0002: One of the operands of idiv is NaN or the first operand is (-)INF'
+				);
+			}
+			if (Number.isFinite(castA.value) && !Number.isFinite(castB.value)) {
+				return createAtomicValue(0, ValueType.XSINTEGER);
+			}
+			return createAtomicValue(fun(castA.value, castB.value), ValueType.XSINTEGER);
+		},
+		ValueType.XSINTEGER,
+	];
+}
+
+const operatorsByTypingKey: Record<string, [BinaryEvaluationFunction, ValueType]> = Object.create(
+	null
+);
+
+export function getBinaryPrefabOperator(
+	leftType: ValueType,
+	rightType: ValueType,
+	operator: string
+) {
+	const typingKey = `${leftType}~${rightType}~${operator}`;
+
+	let prefabOperator = operatorsByTypingKey[typingKey];
+	if (!prefabOperator) {
+		prefabOperator = operatorsByTypingKey[typingKey] = generateBinaryOperatorFunction(
+			operator,
+			leftType,
+			rightType
+		);
+	}
+
+	return prefabOperator;
+}
 
 class BinaryOperator extends Expression {
 	private _evaluateFunction?: BinaryEvaluationFunction;
@@ -510,19 +403,11 @@ class BinaryOperator extends Expression {
 					);
 				}
 
-				// TODO: Find a more errorproof solution, hash collisions can occur here
-				const typingKey = `${firstValue.type}~${secondValue.type}~${this._operator}`;
-				let prefabOperator = operatorsByTypingKey[typingKey];
-				if (!prefabOperator) {
-					prefabOperator = operatorsByTypingKey[
-						typingKey
-					] = generateBinaryOperatorFunction(
-						this._operator,
-						firstValue.type,
-						secondValue.type,
-						this.type
-					);
-				}
+				const prefabOperator = getBinaryPrefabOperator(
+					firstValue.type,
+					secondValue.type,
+					this._operator
+				)[0];
 
 				return sequenceFactory.singleton(prefabOperator(firstValue, secondValue));
 			});
